Adding experimental MultiPlatform support to CMake for Windows and VS 2012 and above.
By setting the CMAKE_VS_EFFECTIVE_PLATFORMS variable, the user is able to
specify that multiple platforms should be produced for the same project.

The supported platforms are Win32, ARM and x64 (not applicable to Windows Phone)
and should be specified as a delimited string Win32;ARM;x64 for example.